+++ src/tests.rs
@@ -123,6 +123,116 @@ macro_rules! test {
                 assert_eq!(1.0, m1.get(kilogram));
             }
 
+            #[cfg(feature = "std")]
+            #[test]
+            fn floor() {
+                let l1 = TLength::new::<kilometer>(3.9999);
+                let l2 = TLength::new::<kilometer>(3.0001);
+                let l3 = TLength::new::<meter>(3.9999);
+                let l4 = TLength::new::<meter>(3.0001);
+                let m1 = TMass::new::<kilogram>(3.9999);
+                let m2 = TMass::new::<kilogram>(3.0001);
+
+                assert_eq!(3.0, l1.floor(kilometer).get(kilometer));
+                assert_eq!(3999.0, l1.floor(meter).get(meter));
+                assert_eq!(3.0, l2.floor(kilometer).get(kilometer));
+                assert_eq!(3000.0, l2.floor(meter).get(meter));
+                assert_eq!(0.0, l3.floor(kilometer).get(kilometer));
+                assert_eq!(3.0, l3.floor(meter).get(meter));
+                assert_eq!(0.0, l4.floor(kilometer).get(kilometer));
+                assert_eq!(3.0, l4.floor(meter).get(meter));
+                assert_eq!(3.0, m1.floor(kilogram).get(kilogram));
+                assert_eq!(3.0, m2.floor(kilogram).get(kilogram));
+            }
+
+            #[cfg(feature = "std")]
+            #[test]
+            fn ceil() {
+                let l1 = TLength::new::<kilometer>(3.9999);
+                let l2 = TLength::new::<kilometer>(3.0001);
+                let l3 = TLength::new::<meter>(3.9999);
+                let l4 = TLength::new::<meter>(3.0001);
+                let m1 = TMass::new::<kilogram>(3.9999);
+                let m2 = TMass::new::<kilogram>(3.0001);
+
+                assert_eq!(4.0, l1.ceil(kilometer).get(kilometer));
+                assert_eq!(4000.0, l1.ceil(meter).get(meter));
+                assert_eq!(4.0, l2.ceil(kilometer).get(kilometer));
+                assert_eq!(3001.0, l2.ceil(meter).get(meter));
+                assert_eq!(1.0, l3.ceil(kilometer).get(kilometer));
+                assert_eq!(4.0, l3.ceil(meter).get(meter));
+                assert_eq!(1.0, l4.ceil(kilometer).get(kilometer));
+                assert_eq!(4.0, l4.ceil(meter).get(meter));
+                assert_eq!(4.0, m1.ceil(kilogram).get(kilogram));
+                assert_eq!(4.0, m2.ceil(kilogram).get(kilogram));
+            }
+
+            #[cfg(feature = "std")]
+            #[test]
+            fn round() {
+                let l1 = TLength::new::<kilometer>(3.3);
+                let l2 = TLength::new::<kilometer>(3.5);
+                let l3 = TLength::new::<meter>(3.3);
+                let l4 = TLength::new::<meter>(3.5);
+                let m1 = TMass::new::<kilogram>(3.3);
+                let m2 = TMass::new::<kilogram>(3.5);
+
+                assert_eq!(3.0, l1.round(kilometer).get(kilometer));
+                assert_eq!(3300.0, l1.round(meter).get(meter));
+                assert_eq!(4.0, l2.round(kilometer).get(kilometer));
+                assert_eq!(3500.0, l2.round(meter).get(meter));
+                assert_eq!(0.0, l3.round(kilometer).get(kilometer));
+                assert_eq!(3.0, l3.round(meter).get(meter));
+                assert_eq!(0.0, l4.round(kilometer).get(kilometer));
+                assert_eq!(4.0, l4.round(meter).get(meter));
+                assert_eq!(3.0, m1.round(kilogram).get(kilogram));
+                assert_eq!(4.0, m2.round(kilogram).get(kilogram));
+            }
+
+            #[cfg(feature = "std")]
+            #[test]
+            fn trunc() {
+                let l1 = TLength::new::<kilometer>(3.3);
+                let l2 = TLength::new::<kilometer>(3.5);
+                let l3 = TLength::new::<meter>(3.3);
+                let l4 = TLength::new::<meter>(3.5);
+                let m1 = TMass::new::<kilogram>(3.3);
+                let m2 = TMass::new::<kilogram>(3.5);
+
+                assert_eq!(3.0, l1.trunc(kilometer).get(kilometer));
+                assert_eq!(3300.0, l1.trunc(meter).get(meter));
+                assert_eq!(3.0, l2.trunc(kilometer).get(kilometer));
+                assert_eq!(3500.0, l2.trunc(meter).get(meter));
+                assert_eq!(0.0, l3.trunc(kilometer).get(kilometer));
+                assert_eq!(3.0, l3.trunc(meter).get(meter));
+                assert_eq!(0.0, l4.trunc(kilometer).get(kilometer));
+                assert_eq!(3.0, l4.trunc(meter).get(meter));
+                assert_eq!(3.0, m1.trunc(kilogram).get(kilogram));
+                assert_eq!(3.0, m2.trunc(kilogram).get(kilogram));
+            }
+
+            #[cfg(feature = "std")]
+            #[test]
+            fn fract() {
+                let l1 = TLength::new::<kilometer>(3.3);
+                let l2 = TLength::new::<kilometer>(3.5);
+                let l3 = TLength::new::<meter>(3.3);
+                let l4 = TLength::new::<meter>(3.5);
+                let m1 = TMass::new::<kilogram>(3.3);
+                let m2 = TMass::new::<kilogram>(3.5);
+
+                ulps_eq!(0.3, l1.fract(kilometer).get(kilometer), epsilon = EPSILON);
+                ulps_eq!(0.0, l1.fract(meter).get(meter), epsilon = EPSILON);
+                ulps_eq!(0.0, l2.fract(kilometer).get(kilometer), epsilon = EPSILON);
+                ulps_eq!(0.5, l2.fract(meter).get(meter), epsilon = EPSILON);
+                ulps_eq!(0.0033, l3.fract(kilometer).get(kilometer), epsilon = EPSILON);
+                ulps_eq!(0.3, l3.fract(meter).get(meter), epsilon = EPSILON);
+                ulps_eq!(0.00035, l4.fract(kilometer).get(kilometer), epsilon = EPSILON);
+                ulps_eq!(0.5, l4.fract(meter).get(meter), epsilon = EPSILON);
+                ulps_eq!(0.3, m1.fract(kilogram).get(kilogram), epsilon = EPSILON);
+                ulps_eq!(0.5, m2.fract(kilogram).get(kilogram), epsilon = EPSILON);
+            }
+
             #[test]
             fn conversion() {
                 assert_eq!(1000.0, kilometer::conversion());
@@ -432,6 +542,116 @@ macro_rules! test {
                 ulps_eq!(1.0, l1.get(kilometer));
                 ulps_eq!(0.001, l2.get(kilometer));
                 ulps_eq!(1.0, m1.get(kilogram));
+            }
+
+            #[cfg(feature = "std")]
+            #[test]
+            fn floor() {
+                let l1 = k::TLength::new::<kilometer>(3.9999);
+                let l2 = k::TLength::new::<kilometer>(3.0001);
+                let l3 = k::TLength::new::<meter>(3.9999);
+                let l4 = k::TLength::new::<meter>(3.0001);
+                let m1 = k::TMass::new::<kilogram>(3.9999);
+                let m2 = k::TMass::new::<kilogram>(3.0001);
+
+                assert_eq!(3.0, l1.floor(kilometer).get(kilometer));
+                assert_eq!(3999.0, l1.floor(meter).get(meter));
+                assert_eq!(3.0, l2.floor(kilometer).get(kilometer));
+                assert_eq!(3000.0, l2.floor(meter).get(meter));
+                assert_eq!(0.0, l3.floor(kilometer).get(kilometer));
+                assert_eq!(3.0, l3.floor(meter).get(meter));
+                assert_eq!(0.0, l4.floor(kilometer).get(kilometer));
+                assert_eq!(3.0, l4.floor(meter).get(meter));
+                assert_eq!(3.0, m1.floor(kilogram).get(kilogram));
+                assert_eq!(3.0, m2.floor(kilogram).get(kilogram));
+            }
+
+            #[cfg(feature = "std")]
+            #[test]
+            fn ceil() {
+                let l1 = k::TLength::new::<kilometer>(3.9999);
+                let l2 = k::TLength::new::<kilometer>(3.0001);
+                let l3 = k::TLength::new::<meter>(3.9999);
+                let l4 = k::TLength::new::<meter>(3.0001);
+                let m1 = k::TMass::new::<kilogram>(3.9999);
+                let m2 = k::TMass::new::<kilogram>(3.0001);
+
+                assert_eq!(4.0, l1.ceil(kilometer).get(kilometer));
+                assert_eq!(4000.0, l1.ceil(meter).get(meter));
+                assert_eq!(4.0, l2.ceil(kilometer).get(kilometer));
+                assert_eq!(3001.0, l2.ceil(meter).get(meter));
+                assert_eq!(1.0, l3.ceil(kilometer).get(kilometer));
+                assert_eq!(4.0, l3.ceil(meter).get(meter));
+                assert_eq!(1.0, l4.ceil(kilometer).get(kilometer));
+                assert_eq!(4.0, l4.ceil(meter).get(meter));
+                assert_eq!(4.0, m1.ceil(kilogram).get(kilogram));
+                assert_eq!(4.0, m2.ceil(kilogram).get(kilogram));
+            }
+
+            #[cfg(feature = "std")]
+            #[test]
+            fn round() {
+                let l1 = k::TLength::new::<kilometer>(3.3);
+                let l2 = k::TLength::new::<kilometer>(3.5);
+                let l3 = k::TLength::new::<meter>(3.3);
+                let l4 = k::TLength::new::<meter>(3.5);
+                let m1 = k::TMass::new::<kilogram>(3.3);
+                let m2 = k::TMass::new::<kilogram>(3.5);
+
+                assert_eq!(3.0, l1.round(kilometer).get(kilometer));
+                assert_eq!(3300.0, l1.round(meter).get(meter));
+                assert_eq!(4.0, l2.round(kilometer).get(kilometer));
+                assert_eq!(3500.0, l2.round(meter).get(meter));
+                assert_eq!(0.0, l3.round(kilometer).get(kilometer));
+                assert_eq!(3.0, l3.round(meter).get(meter));
+                assert_eq!(0.0, l4.round(kilometer).get(kilometer));
+                assert_eq!(4.0, l4.round(meter).get(meter));
+                assert_eq!(3.0, m1.round(kilogram).get(kilogram));
+                assert_eq!(4.0, m2.round(kilogram).get(kilogram));
+            }
+
+            #[cfg(feature = "std")]
+            #[test]
+            fn trunc() {
+                let l1 = k::TLength::new::<kilometer>(3.3);
+                let l2 = k::TLength::new::<kilometer>(3.5);
+                let l3 = k::TLength::new::<meter>(3.3);
+                let l4 = k::TLength::new::<meter>(3.5);
+                let m1 = k::TMass::new::<kilogram>(3.3);
+                let m2 = k::TMass::new::<kilogram>(3.5);
+
+                assert_eq!(3.0, l1.trunc(kilometer).get(kilometer));
+                assert_eq!(3300.0, l1.trunc(meter).get(meter));
+                assert_eq!(3.0, l2.trunc(kilometer).get(kilometer));
+                assert_eq!(3500.0, l2.trunc(meter).get(meter));
+                assert_eq!(0.0, l3.trunc(kilometer).get(kilometer));
+                assert_eq!(3.0, l3.trunc(meter).get(meter));
+                assert_eq!(0.0, l4.trunc(kilometer).get(kilometer));
+                assert_eq!(3.0, l4.trunc(meter).get(meter));
+                assert_eq!(3.0, m1.trunc(kilogram).get(kilogram));
+                assert_eq!(3.0, m2.trunc(kilogram).get(kilogram));
+            }
+
+            #[cfg(feature = "std")]
+            #[test]
+            fn fract() {
+                let l1 = k::TLength::new::<kilometer>(3.3);
+                let l2 = k::TLength::new::<kilometer>(3.5);
+                let l3 = k::TLength::new::<meter>(3.3);
+                let l4 = k::TLength::new::<meter>(3.5);
+                let m1 = k::TMass::new::<kilogram>(3.3);
+                let m2 = k::TMass::new::<kilogram>(3.5);
+
+                ulps_eq!(0.3, l1.fract(kilometer).get(kilometer), epsilon = EPSILON);
+                ulps_eq!(0.0, l1.fract(meter).get(meter), epsilon = EPSILON);
+                ulps_eq!(0.0, l2.fract(kilometer).get(kilometer), epsilon = EPSILON);
+                ulps_eq!(0.5, l2.fract(meter).get(meter), epsilon = EPSILON);
+                ulps_eq!(0.0033, l3.fract(kilometer).get(kilometer), epsilon = EPSILON);
+                ulps_eq!(0.3, l3.fract(meter).get(meter), epsilon = EPSILON);
+                ulps_eq!(0.00035, l4.fract(kilometer).get(kilometer), epsilon = EPSILON);
+                ulps_eq!(0.5, l4.fract(meter).get(meter), epsilon = EPSILON);
+                ulps_eq!(0.3, m1.fract(kilogram).get(kilogram), epsilon = EPSILON);
+                ulps_eq!(0.5, m2.fract(kilogram).get(kilogram), epsilon = EPSILON);
             }
 
             quickcheck! {